pyproject.toml
[project]
name = "automatic-tweet-generator"
version = "0.1.0"
description = "Automated AI/Data Science tweet generator and publisher"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0",
    "google-genai>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "tweepy>=4.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "flake8>=7.0",
]

[tool.black]
line-length = 140

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
